more refactoring of common payment libs
