+++ libs/payments/streams.go
@@ -1,14 +1,21 @@
 package payments
 
 const (
-	PreparePrefix  = "prepare-"
-	SubmitPrefix   = "submit-"
+	// PreparePrefix is the prefix for streams dealing with prepare events
+	PreparePrefix = "prepare-"
+	// SubmitPrefix is the prefix for streams dealing with submit events
+	SubmitPrefix = "submit-"
+	// ResponseSuffix is the suffix for streams dealing with responses
 	ResponseSuffix = "-response"
 )
 
 var (
-	PrepareConfigStream        = PreparePrefix + "config"
+	// PrepareConfigStream is the stream for configuration events for new settlement prepare streams
+	PrepareConfigStream = PreparePrefix + "config"
+	// PrepareConfigConsumerGroup is the consumergroup for the prepare config stream
 	PrepareConfigConsumerGroup = PrepareConfigStream + "-wg"
-	SubmitConfigStream         = SubmitPrefix + "config"
-	SubmitConfigConsumerGroup  = SubmitConfigStream + "-wg"
+	// SubmitConfigStream is the stream for configuration events for new settlement submit streams
+	SubmitConfigStream = SubmitPrefix + "config"
+	// SubmitConfigConsumerGroup is the consumergroup for the submit config stream
+	SubmitConfigConsumerGroup = SubmitConfigStream + "-wg"
 )
